Release 1.03.03
git-svn-id: http://svn.code.sf.net/p/xmlrpc-c/code@435 98333e67-4a24-44d7-a75c-e53540dd3050

+++ branches/release_1_03/xmlrpc-c/lib/abyss/src/conf.c
@@ -306,7 +306,7 @@ abyss_bool ConfReadServerFile(const char *filename,TServer *srv)
                             TraceExit("Unknown user '%s'",p);
         
                         srv->uid=pwd->pw_uid;
-                        if (srv->gid==(-1))
+                        if ((int)srv->gid==(-1))
                             srv->gid=pwd->pw_gid;
                     };
 #else
